# specs/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Gilgamesh Platform API
  version: 1.0.0
  summary: Multi-tenant QA platform API — agents plan, author and run tests.
  description: |
    Gilgamesh v1 HTTP contract. Multi-tenant QA platform where 11 deity-agents plan, author and
    run tests, collaborate on a DAG orchestration canvas, and keep results in-app.

    This document is the **API contract** expansion of the Foundation Keystone
    (`specs/_keystone/foundation-vocabulary.md`). Entity names, field names, enum values, agent
    slot ids and resource paths are taken **verbatim** from the keystone (§1, §2, §3, §5, §6).

    ## Conventions (apply to EVERY operation)
    - **Base path**: all paths below are relative to the server URL `/api/v1`.
    - **Tenant isolation (primordial)**: every request resolves the tenant from the authenticated
      session → `orgId`. Every list/detail query is filtered by that `orgId` (row-level isolation).
      `orgId` is **never** accepted in a request body — a client-supplied `orgId` would be a
      cross-tenant injection vector. For `/orgs/{orgId}/...` paths the path `orgId` is additionally
      checked against the caller's memberships. A **non-member** receives `404` (the resource's
      existence is not leaked); a member with an **insufficient role** receives `403`.
    - **Auth**: a `__Host-`-prefixed, httpOnly + Secure + SameSite=Lax session cookie
...
      `Domain` attribute. See `securitySchemes`. All operations require it except
...

...
  contact:
...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /orgs/{orgId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    get:
      tags: [orgs]
      operationId: getOrg
      summary: Get an org
      description: |
        Caller must be a member of `{orgId}`. A **non-member receives `404`** (the org's existence is
        not leaked — slice-1 §10.2); `403` is reserved for a member with an insufficient role.
      responses:
        '200':
          description: Org detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrgView' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/ProblemDefault' }
    patch:
      tags: [orgs]
      operationId: updateOrg
      summary: Update an org
      description: "RBAC: OWNER or ADMIN. A non-member receives `404`; insufficient role → `403`."
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrgUpdate' }
            example: { name: OmniPizza International }
      responses:
        '200':
          description: Updated org.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrgView' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/ProblemDefault' }

  # =====================================================================================
  # MEMBERS  (FULL DETAIL)
  # =====================================================================================
...
  /orgs/{orgId}/agents:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    get:
      tags: [agents]
      operationId: listOrgAgents
      summary: Org agent catalog (the 11)
      description: |
        Returns the per-Org agent catalog seeded from the canonical roster (keystone §3): 11 agents,
        unique by `slot`. Each item includes the per-role tool options (`toolOptions`, first = default).
        This catalog is effectively static (it changes only on Org seeding), so the response carries an
        `ETag` + `Cache-Control`; clients SHOULD cache it (e.g. React Query/SWR keyed on `orgId`) and
        re-validate with `If-None-Match` rather than refetching on every Agent-room mount (`304`).
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The 11 agents.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AgentView' }
              example:
                data:
                  - id: 0190f3c0-0001-7000-8000-000000000001
                    orgId: 0190f3a2-aaaa-7bbb-8ccc-001122334455
                    slot: lead
                    deityName: Zeus
                    role: QA Lead
                    family: proceso
                    glyph: ZE
                    culture: Grecia
                    defaultTool: Helix Core
                    toolOptions: [Helix Core]
                    createdAt: "2026-06-01T09:05:00Z"
                  - id: 0190f3c0-0001-7000-8000-000000000004
                    orgId: 0190f3a2-aaaa-7bbb-8ccc-001122334455
                    slot: web
                    deityName: Quetzalcóatl
                    role: Web
                    family: ui
                    glyph: QC
                    culture: Azteca
                    defaultTool: Playwright
                    toolOptions: [Playwright, Cypress]
                    createdAt: "2026-06-01T09:05:00Z"
        '304': { $ref: '#/components/responses/NotModified' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/ProblemDefault' }
  /projects/{id}/agents:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [agents]
      operationId: listProjectAgents
      summary: Project agents (catalog + tool binding + runtime status)
      description: |
        Returns each of the 11 agents joined with its per-Project `ToolBinding` (selected tool +
        enabled/awake) and a derived `runtimeStatus` (IDLE if !enabled; BUSY if currently in a running
        RunNode; else ACTIVE). Resolved in **one** query (join, no N+1). Drives the Agent room.
        Changes only on wake/sleep/tool-select, so the response carries an `ETag` + `Cache-Control`;
        clients SHOULD cache it and re-validate with `If-None-Match` (`304`) rather than refetch on
        every mount.
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Project agents.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectAgentView' }
              example:
                data:
                  - agent:
                      id: 0190f3c0-0001-7000-8000-000000000004
                      orgId: 0190f3a2-aaaa-7bbb-8ccc-001122334455
                      slot: web
                      deityName: Quetzalcóatl
                      role: Web
                      family: ui
                      glyph: QC
                      culture: Azteca
                      defaultTool: Playwright
                      toolOptions: [Playwright, Cypress]
                      createdAt: "2026-06-01T09:05:00Z"
                    toolBinding:
                      id: 0190f3d0-0001-7000-8000-000000000004
                      orgId: 0190f3a2-aaaa-7bbb-8ccc-001122334455
                      projectId: 0190f3b0-1111-7222-8333-444455556666
                      agentId: 0190f3c0-0001-7000-8000-000000000004
                      tool: Playwright
                      enabled: true
                      updatedAt: "2026-06-25T19:00:02Z"
                    runtimeStatus: ACTIVE
        '304': { $ref: '#/components/responses/NotModified' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/ProblemDefault' }
  /projects/{id}/agents/{slot}:
    parameters:
      - $ref: '#/components/parameters/PathId'
      - $ref: '#/components/parameters/Slot'
    patch:
      tags: [agents]
      operationId: updateProjectAgent
      summary: Update an agent's tool binding (Strategy) / wake-sleep
      description: |
        Upserts the per-Project `ToolBinding` for the agent in `{slot}`: select the active `tool`
        (must be one of the agent's role tool options) and/or toggle `enabled` (awake/asleep).
        RBAC: MEMBER+.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ToolBindingUpdate' }
            example: { tool: Cypress, enabled: true }
      responses:
        '200':
          description: Updated project agent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectAgentView' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/ProblemDefault' }
  /projects/{id}/agents/wake-all:
    parameters:
      - $ref: '#/components/parameters/PathId'
    post:
      tags: [agents]
      operationId: wakeAllProjectAgents
      summary: "Awaken team — enable all idle agents"
      description: |
        Sets `enabled = true` on every agent's `ToolBinding` for this project (the "Awaken team"
        action). Returns the full updated project-agent list. Idempotent. RBAC: MEMBER+.
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        '200':
          description: All agents awakened.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectAgentView' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/ProblemDefault' }

  # =====================================================================================
  # SLICES  (OUTLINE — TODO: expand)
  # =====================================================================================
...
  /orgs/{orgId}/subscription:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    get:
      tags: [subscription]
      operationId: getSubscription
      summary: Get the org subscription (plan, seats, quota)
      description: |
        TODO: expand. Mock billing behind the PaymentProvider port. Tenant-checked. **RBAC: ADMIN or
        OWNER** (least privilege — full billing fields `plan`/`seats`/`providerCustomerId` are not
        exposed to `MEMBER`/`VIEWER`). Deny-by-default: non-member → `404`, insufficient role → `403`.
      responses:
        '200':
          description: Subscription.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/ProblemDefault' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-gg_session
      description: |
        `__Host-`-prefixed, httpOnly + Secure + SameSite=Lax session cookie issued by `/auth/login` or
        `/auth/register` (slice 1, local auth). The `__Host-` prefix structurally forces `Secure`,
        `Path=/`, and no `Domain`. The session binds the user and the active `orgId` used for tenant
        resolution. Swappable to OIDC later behind the `IdentityProvider` port without contract change.
        **Cookie auth alone is not sufficient for unsafe methods** — they additionally require the
        `csrfToken` double-submit header (see below).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit CSRF token. Required on every **unsafe** method (`POST`/`PATCH`/`PUT`/`DELETE`)
        whenever the request is authenticated by the session cookie. Its value MUST equal the
        non-HttpOnly `csrf` cookie set alongside the session cookie. A missing or mismatched token →
        `403` (`code: CSRF_FAILED`). Not required for safe methods or the public `/auth/*` endpoints.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      description: Org (root tenant) id. Must be one of the caller's memberships.
      schema: { type: string, format: uuid }
    PathId:
      name: id
      in: path
      required: true
      description: Resource id (UUID v7) of the resource named by the path.
      schema: { type: string, format: uuid }
    Slot:
      name: slot
      in: path
      required: true
      description: Agent slot key (keystone §3).
      schema: { $ref: '#/components/schemas/AgentSlot' }
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: Conditional GET — pass the previously returned `ETag`; a still-fresh resource yields `304`.
      schema: { type: string }
  headers:
    ETag:
      description: Entity tag for the representation; use with `If-None-Match` for conditional GETs.
      schema: { type: string }
    CacheControl:
      description: Client-cache directive (e.g. `private, max-age=30, must-revalidate`).
      schema: { type: string }
    RateLimitLimit:
      description: Request quota for the current window.
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests remaining in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Unix epoch seconds when the window resets.
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: "https://gilgamesh.example/problems/unauthorized"
            title: Unauthorized
            status: 401
            code: UNAUTHENTICATED
            detail: A valid session is required.
    Forbidden:
      description: Authenticated but not permitted (RBAC or cross-tenant).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: "https://gilgamesh.example/problems/forbidden"
            title: Forbidden
            status: 403
            code: FORBIDDEN
            detail: Your role does not allow this action.
...
    NotFound:
      description: Resource not found (or not visible to the caller's tenant).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: "https://gilgamesh.example/problems/not-found"
            title: Not Found
            status: 404
            code: NOT_FOUND
    NotModified:
      description: Not modified — the client's `If-None-Match` ETag still matches (cached copy is fresh).
...
    Conflict:
      description: State or uniqueness conflict.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: "https://gilgamesh.example/problems/conflict"
            title: Conflict
            status: 409
            code: CONFLICT
            detail: A resource with that slug already exists in this org.
    UnprocessableEntity:
      description: Validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: "https://gilgamesh.example/problems/validation"
            title: Unprocessable Entity
            status: 422
            code: VALIDATION_FAILED
            errors:
              - field: email
                detail: must be a valid email address
...
    ProblemDefault:
      description: Unexpected error (RFC 9457 Problem).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
            code: INTERNAL
  schemas:
...
    AgentSlot:
      type: string
      description: Stable lowercase agent slot keys (keystone §3).
      enum: [lead, arch, manual, web, api, android, ios, perf, visual, sec, a11y]
    AgentFamily:
      type: string
      description: Stable lowercase agent family keys.
      enum: [proceso, ui, backend, guardian]
    AgentRuntimeStatus:
      type: string
      description: |
        Derived: IDLE if not enabled; BUSY if the agent is in a currently running RunNode; else ACTIVE.
      enum: [ACTIVE, BUSY, IDLE]
...
    Plan:
      type: string
      enum: [TEAM, PRO, ENTERPRISE]
    BillingCycle:
      type: string
      enum: [MONTHLY, ANNUAL]
    SubscriptionStatus:
      type: string
      enum: [TRIALING, ACTIVE, PAST_DUE, CANCELED]
...
    Org:
      type: object
      description: Root tenant.
      required: [id, name, slug, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, description: UUID v7 }
        name: { type: string }
        slug: { type: string, description: Unique org slug. }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Agent:
      type: object
      description: Per-Org agent catalog entry (one of the 11). Unique(orgId,slot). Seeded from roster §3.
      required: [id, orgId, slot, deityName, role, family, glyph, culture, defaultTool, createdAt]
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        slot: { $ref: '#/components/schemas/AgentSlot' }
        deityName: { type: string, examples: [Zeus] }
        role: { type: string, description: "Human label, e.g. QA Lead." }
        family: { $ref: '#/components/schemas/AgentFamily' }
        glyph: { type: string, examples: [ZE] }
        culture: { type: string, examples: [Grecia] }
        defaultTool: { type: string, examples: [Helix Core] }
        createdAt: { type: string, format: date-time }
    ToolBinding:
      type: object
      description: Per-Project agent state + tool selection (Strategy). Unique(projectId,agentId).
      required: [id, orgId, projectId, agentId, tool, enabled, updatedAt]
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        agentId: { type: string, format: uuid }
        tool: { type: string, description: Selected tool (∈ the agent's role options). }
        enabled: { type: boolean, description: Awake when true. }
        updatedAt: { type: string, format: date-time }
...
    Subscription:
      type: object
      description: Org subscription (mock billing now). Unique(orgId).
      required: [id, orgId, plan, billingCycle, seats, status, runMinutesQuota, runMinutesUsed]
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        plan: { $ref: '#/components/schemas/Plan' }
        billingCycle: { $ref: '#/components/schemas/BillingCycle' }
        seats: { type: integer }
        status: { $ref: '#/components/schemas/SubscriptionStatus' }
        runMinutesQuota: { type: integer }
        runMinutesUsed: { type: integer }
        providerCustomerId: { type: string }
        providerSubscriptionId: { type: string }
        currentPeriodEnd: { type: string, format: date-time }
...
    OrgView:
      allOf: [ { $ref: '#/components/schemas/Org' } ]
      description: Org response DTO.
...
    AgentView:
      allOf:
        - { $ref: '#/components/schemas/Agent' }
        - type: object
          required: [toolOptions]
          properties:
            toolOptions:
              type: array
              items: { type: string }
              description: Per-role tool options (keystone §3); first element is the default tool.
      description: Agent catalog response DTO (entity + available tool options).
    ToolBindingView:
      allOf: [ { $ref: '#/components/schemas/ToolBinding' } ]
      description: ToolBinding response DTO.
    ProjectAgentView:
      type: object
      description: |
        Composite project-agent view = catalog `Agent` + its per-Project `ToolBinding` + derived
        `runtimeStatus`. Drives the Agent room and the Strategy tool selector.
      required: [agent, runtimeStatus]
      properties:
        agent: { $ref: '#/components/schemas/AgentView' }
        toolBinding:
          oneOf:
            - { $ref: '#/components/schemas/ToolBindingView' }
            - { type: 'null' }
          description: Null until a binding is first created for this project.
        runtimeStatus: { $ref: '#/components/schemas/AgentRuntimeStatus' }
...
    SubscriptionView:
      allOf: [ { $ref: '#/components/schemas/Subscription' } ]
      description: Subscription response DTO.
...
    OrgUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string }
        slug: { type: string }
...
    ToolBindingUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      description: PATCH a project agent — select tool and/or toggle awake.
      properties:
        tool: { type: string, description: Must be one of the agent's role tool options. }
        enabled: { type: boolean }
...
    Problem:
      type: object
      description: RFC 9457 problem details (`application/problem+json`).
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: about:blank, description: A URI identifying the problem type. }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string, description: Stable machine-readable error code. }
        errors:
          type: array
          description: Field-level validation errors (extension member).
          items:
            type: object
            properties:
              field: { type: string }
              detail: { type: string }